mvoe static files and implement them through express, serve html from express, map api result correctly
