+++ app.js
@@ -1,10 +1,15 @@
 'use strict';
 var express = require('express');
 var app = express();
-const lookup = require('country-data').lookup;
 const controllers = require('./controllers/controllersIndex.js');
+const path = require('path');
 
-app.get('/:petitionId', function (req, res) {
+app.use(express.static('public'));
+app.get('/', function (req, res) {
+  res.sendFile(path.join(__dirname + '/countries.html'));
+});
+
+app.get('/api/:petitionId', function (req, res) {
   controllers.petitionController(req.params.petitionId, (result) => {
     res.send(result);
   });
@@ -13,16 +18,3 @@ app.get('/:petitionId', function (req, res) {
 app.listen(3000, function () {
   console.log('Example app listening on port 3000!');
 });
-
-var addCountryCode = (countries) => {
-  var formattedCountries = countries.map((country)=> {
-    country.code = lookupCountry(country.name);
-    console.log(`country: ${country}`);
-    return country;
-  });
-  console.log(formattedCountries);
-  return formattedCountries;
-};
- var lookupCountry = (countryName) => {
-  return lookup.countries({name: countryName})[0] ? lookup.countries({name: countryName})[0].alpha3 : null;
-};
